refractor: Use testDenoXRun method in tests
