+++ src/utils/file.ts
@@ -15,4 +15,17 @@ function readFirstExistingFile(files: string[]): string {
   }
 }
 
-export { readFirstExistingFile };
+async function exists(filename: string): Promise<boolean> {
+  try {
+    await Deno.stat(filename);
+    return true;
+  } catch (error) {
+    if (error instanceof Deno.errors.NotFound) {
+      return false;
+    } else {
+      throw error;
+    }
+  }
+}
+
+export { readFirstExistingFile, exists };
